WIP change the way project users management works

+++ application/src/Service/UserProjectStatusManager.php
@@ -7,6 +7,7 @@ use App\Entity\User;
 use App\Entity\Project;
 use App\Entity\UserProject;
 use App\Service\AppEnums;
+use App\Service\FlashManager;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
 use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface;
@@ -16,12 +17,14 @@ class UserProjectStatusManager
     protected $em;
     protected $authChecker;
     protected $user;
+    protected $fm;
 
-    public function __construct(EntityManagerInterface $em, AuthorizationCheckerInterface $authChecker, TokenStorageInterface $tokenStorage)
+    public function __construct(EntityManagerInterface $em, AuthorizationCheckerInterface $authChecker, TokenStorageInterface $tokenStorage, FlashManager $fm)
     {
         $this->em = $em;
         $this->authChecker = $authChecker;
         $this->user = $tokenStorage->getToken()->getUser();
+        $this->fm = $fm;
     }
 
     public function create(Project $project)
@@ -44,19 +47,45 @@ class UserProjectStatusManager
     public function toggle(UserProjectStatus $userProjectStatus)
     {
         $enabled = $userProjectStatus->getEnabled();
-        $userProjectStatus->setEnabled(!$enabled);
+        $statusName = $userProjectStatus->getStatus()->getName();
 
-        $this->em->persist($userProjectStatus);
-        $this->em->flush();
+
+        $canToggle = ($enabled && $statusName == AppEnums::USER_STATUS_MANAGER_NAME)
+          ? $this->hasManager($userProjectStatus)
+          : true;
+
+
+        if ($canToggle) {
+            $userProjectStatus->setEnabled(!$enabled);
+
+            $this->em->persist($userProjectStatus);
+            $this->em->flush();
+        }
 
         return $userProjectStatus;
     }
 
     public function remove(UserProjectStatus $userProjectStatus)
     {
-        $this->em->remove($userProjectStatus);
-        $this->em->flush();
+        if ($this->hasManager($userProjectStatus)) {
+            $this->em->remove($userProjectStatus);
+            $this->em->flush();
+        }
 
-        return $userProjectStatus;
+        return;
+    }
+
+    private function hasManager(UserProjectStatus $userProjectStatus)
+    {
+        $project = $userProjectStatus->getProject();
+
+        $count = $this->em->getRepository("App:UserProjectStatus")->countManagerByProject($project);
+
+        if ($count < 2) {
+            $this->fm->add('warning', 'need_at_least_one_manager');
+            return false;
+        }
+
+        return true;
     }
 }
